refactor(calculations): convert financial data class to plain type
The goal is to simplify the interface of the calculation. The current
class has an empty constructor and all properties get overwritten later
on. This is not a great approach. The defaults used for the class
properties are not really necessary. Not having them, provides more
transparent usage. In terms of usage convenience there are possibilities
to further improve this type in future.

In functional style, the data is marked as readonly. But unfortunately
that is not completely possible. This is due to the flaws in the
calculation we are already aware of. It makes this fully visible and
more obvious to recognize now.

The cloning is still necessary and a pita due to the usage of Big.js
values.