refactor(calculation): release methods of MischEkTaetigkeit class
We strive to more simple data representations and logic that reads and
evaluates their data. Therefore, the lonely method of the
MischEkTaetigkeit class has been resolved. It does not include any fancy
domain logic and is very straight forward to apply where needed. It can
actually be improved for the individual use-case and "cached".