refactor(calculation): simplify MischEkTaetigkeit class
The property `bemessungsZeitraumMonate` is basically a list of flags. It
marks for each month if the Tätigkeit took place in a month (by index)
or not. As for all flags, if the value is absent, its equal to the
negative case. For a list with unsafe indexes, this also just means that
if there is no entry for this month index there was no Tätigkeit. Which
in turn means that an empty array is fine too and that there is no need
for having to initialize it strictly on construction for 12 months.