refactor(calculation): release methods of ErwerbsZeitraumLebensMonat
The goal is to simplify the classes as simple data structures where
possible. Following a more functional approach.
One of the class its methods was simply unused. The other used at
a single location for a single call. Therefore this method has been
moved as a plain function int the respective module. Tests have been
moved too, though this finally starts to use in-source tests also for
the calculations part of the application.
Notice that despite such refactoring try to focus on structure, the
implementation of the moved function was rewritten too, as it was just
suboptimal.