feat(planer): accessibility description for Lebensmonat summary
The description of a Lebensmonat, especially important for the collapsed
state, was broken. The description of the summary element was referring
to an element within the content via an identifier. Due to performance
optimization, the content has been changed to render only when the
Lebensmonat is expanded (before it was just not visible). Thereby, the
identifier resolves to nothing and the description is lacking
information.
To resolve this issue, this rather weird relation was discarded
(introduced by myself). The summary is already responsible to generate
a full description of the Lebensmonat by itself, despite the Zeitraum.
So this was just a natural extension. In result, the whole component
relationships could be simplified and made more stable.

The Zeitraum label component does no more need the extra accessibility
optimizations. It was only for this one usage. The other usages of the
component are more simple. The component is still helpful, even though
its small responsibility, but high customizability.

The diff of the commit looks suspiciously big. A huge part of it is just
indentations and formatting. ;-)