refactor(calculation): remove Brutto- and Nettoeinkommen from model
Having an explicit and expressive domain model is highly desirable.
Especially value object play an important role. Within the calculations
domain model, the type `Einkommen` is alias by `BruttoEinkommen` and
`NettoEinkommen`. Because it is an alias, the only advantage it can
provide is expressiveness. But every location it is used, the property
name, function name, ... is already expressing this information. So it
doesn't add anything.
These types do not add any strong separation and operation segregation,
which is quite typical for domain value object (like unit rich types).
This is neither implemented, nor necessary. In fact, everywhere a gross
or net income is needed, a plain income value is instantiated. Type
alias in TypeScript are very week.
As they add no value, they got removed for simplification.

This commit is in the context to simplify the whole calculation domain.
The domain itself is quite complicated. The code should not make it
unnecessarily more hard to understand. Though, this is a tiny nitpick
thing here, it adds up with other refactoring.