feat: remove segment filter in metrics-sync
Segment filters are not suitable for filtering data as
they do not return accurate results.

This issue is evident when examining Matomo events
from December 24, 2024. Without the segment filter,
6 ease events are recorded, each associated with a
proper user session. When filtering the page by the
“planer” segment, one event is excluded, resulting
in only 5 ease events, despite the fact that the event
was created in the “planer” segment.

Related Issue: https://github.com/matomo-org/matomo/issues/15438