refactor(calculation): in-source child utilities
The former child utilities have left two methods remaining. These get
used by a single module only. Therefore, it is more lean to in-source it
to the algorithm where it is directly used. Including the tests. This
also avoids any exposing of the logic, avoiding dependencies.
Furthermore, this helps to make the whole calculation code more terse
and simpler to understand, having less to look at left and right.

The former implementation was rather clunky and awkward. A much shorter
and "elegant" solution has been used and tested as replacement.