ci: reschedule metrics sync to run at 4am utc
The previous schedule at the end of the day risked missing late events. Since
the data doesn't need to be available in the middle of the night, 4am provides
a better balance. This timing ensures a) complete data from the previous day.
and b) availability before the start of a new working day, even accounting
for time zones and daylight saving time.