refactor(calculation): convert calculation parameters to plain numbers
The general goal is to get rid of Big.js within the calculation feature.
It brings no extra value and only produces complexity and less readable
code.
As a first simple step, the static calculation parameters have been
converted from Big.js numbers to plain JavaScript numbers. On the way of
fixing their usage, surrounding code has started to be refactored
without using Big.js.
The idea is to continue this process to do this big refactoring step by
step.

After all, there is some difference noticeable in the calculated
numbers. Every few thousand test runs, a value diverges by one cent.
This loss of precision (if it actually is any) is more than acceptable
and was addressed in the test code.