test(calculation): improve equality checks for numbers
There are a few deviations in the test numbers when comparing results
with expected values from an Excel sheet. These seem to origin in
precision issues between the numeric systems for floating numbers in
use.
In the past, this was addressed by some rounding of values and comparing
a value against a range of 5 cent values (e.g. 1.0 gets compared with
0.8, 0.9, 1.0, 1.1, 1.2). In case there is no rounding on the cent
decimal in place, having an expected result of 1.0999999 will fail.
A more solid approach to address this issues is to just use the test
framework expectation statements which support such cases natively.