refactor(calculation): use plain numbers for Mischeinkommen Zwischenergebnis
The general goal is to get rid of Big.js within the calculation feature.
It brings no extra value and only produces complexity and less readable
code.
This commit takes the Zwischenergebnis of Mischeinkommen data type as
a driver for change. Some parts of the algorithm get nicely cleaned up
by this.