refactor(calculation): use plain numbers for ElternGeldPlusErgebnis
The general goal is to get rid of Big.js within the calculation feature.
It brings no extra value and only produces complexity and less readable
code.
The current drive for change was the final result type of the algorithm.
It help to cut of some loosing ends in many parts of the code.
Especially the tax part is no Big.js free.

This also expects all monetary values to match with an error margin in
the refactoring tests. In many cases it is just a matter of how many
million test runs until a rounding issue happens.