fix: some info-i not clickable in Safari
The legend element is special. At least in Safari it doesn't allow to
contain clickable elements. It's not a regular container, although it
works in other browsers.

Absolute positioning seemed to be the only solution, as the legend
element must be the first direct child of the fieldset. And the wrapper
div with position: relative is needed because of the peculiar
relationship between fieldset and legend (the legend sits on the top
border of the fieldset and cannot be moved).