refactor(calculation): erase unnecessary usage of classes
The majority of remaining classes for the calculation are very likely
a remnant of the conversion of Java to JavaScript code. Java does (did?)
not allow functions as first class. They always must be methods of
a class. So especially utility functions become static methods.
To simplify and unclutter the algorithms, all classes have been removed
for the algorithmic part. Classes are still a powerful design approach
when it fits. Here it doesn't. Furthermore it helps to keep the side
effect more under control. It also simplifies the interfaces slightly.
More importantly, having module local functions allows for easy
in-source testing. This is not possible with private class methods.
Only protected methods could be tested then (this practice was not
applied).

This commit makes no attempt in improving any naming or the
expressiveness of function/method names. Everything stays as is for now.
The former wrapping class names did not provide any useful additional
information either. Further refactoring will have this topic as target.