refactor: simplify maximum birthday check
First of all, the tests just mock the system time. This makes the test
data very straight forward and super easy to verify and read.
Unfortunately the testing-library for events in React has a open bug for
mocked time(rs). Therefore there is a little setup overhead that can be
hopefully removed with a future update.

Moreover, the actual implementation has been simplified too. The raw
birthdate three years ago is used for comparison. This does not require
any facy special logic. Though, the whole handling of dates and the
input field is very clunky.