fix(calculations): map correct Programmablaufplan for 2021
This seem like it was a copy-paste mistake when the year 2021 was added
(last). Luckily, we barely ever have users that need such a long ago
year (probably never). So this should have had no impact on production.
Furthermore, the impact on the Elterngeld is minimal too. Anyhow, not
good.