refactor(calculation): wrap-up Big.js removal
This converts the final bits of the calculation code to work without
Big.js. Furthermore it removes the dependencies and some related setup
code. After all, the application is no free of Big.js and uses plain
numbers only. The bundle size also dropped due to this.
There is still and indirect dependency via the tests using the original
algorithm for refactoring property tests.