test(calculations): migrate test specification files to in-source tests
This is our general approach in the code base by now. Though, there are
still a few exceptions like visual components. Anyhow, for modules like
the ones for the calculation features, this fits perfectly. It allows to
reduce the exports, keep the interfaces small, but allow for better
testability at the same time. Plus all the other great advantages of
in-source tests.
Furthermore, a driver of this change was to continue shim down the
calculation feature. It is still cluttered and partially very hard to
understand. Limiting the files/modules to a more meaningful set should
help to reduce the stress when encountering this feature.
Notice that the FIT tests are still in separate specification files. The
future of these tests is uncertain. They are definitely no plain unit
tests, so they should not be put in-source.