refactor(planer): convert Kontingent specification without a class
The intend is to actually never define a specification as a class, but
make use of the extensive interface it provides in a more "functional"
manner. This should improve the readability as the definitions can be
quite terse and easy to read this way. Especially because all the
generic types are derived automatically.
This specification used to be more complex and thereby actually needed
to be a class to fulfill these requirements. But that is no more the
case and could be simplified.