refactor: leverage shouldUnregister to clean up absent fields
The motivation for this change was a bug where the form became non-reactive once the user
submitted it for the first time. We realized that the form was using effects to clean up the
state whenever a dependent field changed. The effect approach was hard to understand, not very
performant, and, in this case, particularly prone to errors.

I discovered the shouldUnregister flag of react-hook-form, which deletes all fields in the form
state when their corresponding components are removed from the DOM. This fulfills our
requirements without any additional code.

To make our code work with shouldUnregister, it was necessary to make 'Elternteil 2'
optional in our domain, which better represents the real world anyway.