fix(planer): domain language for Auswahlmöglichkeiten
Some "technical English" slipped into the Planer domain for the value
objects of Auswahlmöglichkeiten. Formerly, the term `disabled` and
`hint` was used. First of all, these are English. But they are also very
technical, almost having user-interface inputs in mind.
With the renaming to German and more domain native, the "disabled"
flag has switched states. Like true and false work in the opposite way
now. This is because the flag uses a positive name now without negation.
It causes the commit to be not 100% trivial to ensure everything works
still as intended.
There was the intend to use an optional flag. It would have simplified
especially the test data generation as the option could have been
skipped everywhere. Unfortunately with the positive naming this does not
work out anymore. Potentially this is something to revisit and improve
further on.