feat(planer): extend correctness of Partnermonate
The law for Partnermonate is quite complex and detailed. The Planer
can't depict all of them. So it is always a trade off to balance
detailed correctness with simplicity and easy of understanding.
In this context it has been decided to extend the rule. The Planer
domain was extended to be fed with the information if an Elternteil was
erwerbstätig within the Bemessungszeitraum or not. This gets included to
determine if the Partnermonate are available or not.
Notice that there are still missing parts when Partnermonate are truly
applicable by law.