refactor(planer): slim down domain by removing user-interface parts
While developing the domain of the Planer feature, some data structures
and operations didn't feel like they fit. There was no proper domain
terminology for it and it alls was rather user-interface related. In
fact, some parts are just transforming data from domain models to a view
model. Therefore, they got removed from the domain, defining more clear
boundaries and responsibilities.

Due to this, some user-interface components start to just take a plan
and show some information of it in different transformed, assembled,
collected views. This makes them very nice as they are very
self-contained and plug-in like. It also means that the user-interface
service starts to become more slim and focused.

Notice that some ugly things start to happen. The domain defines some
type-safe utilities to act on the domain data structures. These have
been partially exposed under the coat of "domain operations". They
actually delude the clear domain to not expose the internals. Anyhow,
more of such functions just needed to be created to keep this
maintained.
In future, it is the goal to expose a variety of type related utilities
per domain type. Like it is already done for type guards or comparator
functions. This will allow for wide spread adoption of working with the
data of the domain outside the domain in a safe manner. Though, this is
a little tricky, as it requires very generic interfaces to be defined.