refactor(planer): extract Kontingent state from service
In row with recent other changes, some responsibilities in the planer
feature should be toughen. So did the Planer service for the
user-interface grow quite a lot. It got large and complex. And it still
is. But it was cleaned from a bunch of unnecessary "view only" states.
The final piece of this transition was to also remove the Kontingent. It
is similar to the former changes where a component provides basically
just a special view on the Plan data. Just that in this case are the
methods to derive the view from the domain itself. This is because it is
by "coincidence" an important part of the domain itself. But happens to
be valuable for a view component too.

Notice that these changes lead to more standalone and plug-in
components, less prop-drilling. But with cost in performance. Before
this, the service knew on which callbacks it had to update which state.
Now, every time the plan changes, it will cause these components to
re-render and re-calculate. Currently, this isn't an issue. In most
cases the views should just update. It is more simple, easier to test
and more reliable. In worse case, there might be other mechanisms to
improve rendering time (as it is already partially done).