feat(calculations): introduce latest Werbekostenpauschale with history
The main part of this commit is the addition of the new
`werbekostenpauschale` module as an replacement for the `PAUSCH`
constant variable.
The calculation suffers from incorrect results, because a very old
amount is used for the Werbekostenpauschale. So this value needs to be
updated. Anyhow, this amount is known to change over time. And the
correct amount to use depends on the date of birth of the child. To
improve the correctness and to have a complete representation of the
domain, a history of the Werbekostenpauschale has been introduced. It
will then be used to determine which value to use, including the new
amounts that needed to be add. Further details can be found in the code
documentation itself.

The remaining changes are primarily for parameter drilling of the date
of birth. Furthermore some tests had to be adapted to use a date
matching their expectations. The refactoring property tests still work
fine as the inputs are shrunk for the original-rechner anyway.