fix(calculations): remove Lohnsteuer parameter
The entry function to calculate the Elterngeld expects the caller to
define the Lohnsteuerjahr to use. This actually makes no sense. Because
by the regulation of the real world domain, the Lohnsteuerjahr is
determined based on the date of birth of the child. Therefore, this
parameter got removed and swapped. Thanks to the earlier parameter
drilling of the birthday, this was very simple. In the end, it was
basically a swap of both parameters, taking both commits together.

Unfortunately this triggered some testing issues. Before, the birthday
and the tax year were separated input parameter and not required to be
in sync. And some tests actually did not align them properly. In result,
the tests results were not matching. This was unsolvable, because due to
the recent Werbekostenpauschale history, the date of birth has multiple
effects in the calculation for the taxes. Therefore, the was no possible
fix on the input parameters (i.e. the date of birth). As we consider
these tests by now as pure regression tests from old times, the expected
outputs were simply adapted based on the latest calculation runs. The
differences could be traced back to the unmatching Werbekostenpauschale
and tax year.

Furthermore, the refactoring property tests also struggled with this.
The original-rechner still has these separated input parameters, the
production algorithm no more. Therefore, the first input adaption had to
be introduced as an exception (so far). So also the arbitrarily
generated birthdays had to be limited in relation to the supported tax
years of the original-rechner. The latter also had to receive an update
of its Werbekostenpauschale which wasn't matching these tax years.
Luckily, for the two years of birth matching the supported tax years,
the Werbekostenpauschale was constant. Which suits the constant approach
of the original-rechner. Else the years of birth would have had to be
restricted even further.