refactor: replace page scss with tailwind classes and a bit of css
Having a different grid template for a specific breakpoint can
become cumbersome in tailwind, which is why I moved it to a
small css file in the styles directory. We might want to change this
to Tailwind in the future with better tooling, or even migrate our
styles to css modules.