fix: could not parse stylesheet log in test runs
We experience a lot of noise in the test output log because the css parser
used within jsdom fails to parse some (legitimate) cs features. We discussed
different solutions, such as filtering out the logs, switching to happy-dom, and
upgrading the affected libraries in the hope that this would resolve the issue.

Today, in another attempt to fix it, I discovered an option to entirely disable css
in our unit tests. This eliminates the issue without breaking any tests. I believe
this is a good interim solution until the root cause is properly addressed.

https://github.com/jsdom/jsdom/issues/2177
https://digitalservicebund.slack.com/archives/C07T2J3M3B6/p1733129955074979