refactor(calculations): improve low level building blocks around Ersatzleistung
This primarily replaces the former `abstract-algorithm` (which used to
be a class) with functions in the new modules for `ersatzrate` and
`einkommensersatzleistung` (the latter lacks a really good name).
The former implementation was hard to read and depicted with lots of
duplicated code. Furthermore it had zero tests to it.
The new implementation follows pretty straight the law and guidelines.
It uses the exact same terminology and partially also its flow.
Therefore it contains references to the relevant parts of the regulation
that are implemented. Further documentation, as well as the tests should
help any reader to further understand the details.
Moreover, a lot of tests have been added. The in-source tests provide
a mix of simple behavior-driven test cases, parametrized tests with
border values, property tests and tons of examples from the guidelines
(including their appendix) and the so called Dumbatzi-Spickliste. All
together it highly increases the trust in these low level building
blocks of the overall Elterngeld calculation algorithm.

Notice that the usage of the new functions still follows the legacy
code. Unfortunately, the caller logic/algorithm is still pretty broken
and leads to plenty of broken tests when trying to correct it a tiny
bit. This is target for further refactoring. Which likely needs to newly
anticipated testing strategy.