test(calculations): at first iteration of characterization tests
In the context of trying to improve the testing strategy of the calculation
feature, characterization tests have been added. Most of the interesting details
actually received some in-source documentation.
Notice that, while the arbitrary generators look very similar to the ones from
the refactoring properties tests, that these serve a different purpose. They
generate sample data for the exact domain models. While in the refactoring tests
a "higher level" format is chosen that serves old and new domain model. Thereby
they can't be reused.