refactor(atoms): extract styled radio input component
Extract the "special" styling of for inputs of type radio into a new
component. This for formerly applied within the radio group component
directly. Having it extracted from their allows it to be re-used (as we
plan to do).

These low level components should be free of the react-hook-form
dependency. Therefore the new component is a plain text-book React
controlled input. In result, the former approach to use the register
approach by react-hook-form does no more work. So the interface of the
radio group changed a little, affecting all callers.

Moreover, there are very minor improvements that are unrelated to the
topics above (e.g. derive disabled state).