refactor(components): simplify directory structure and imports
This is another follow up to release React component modules from their
folder. In result of past refactoring, many of these folder just contain
the component module file and an index module. That does not add any
value and just clutters the directory structure.

Furthermore, the imports of the components should be more independent
and clean. This a general strategy we try to follow across the
repository. Especially because some part of the application does it,
some not. So all component "levels" use a proper index file now, which
is directly addressed by imports. This helps to blackbox these code
trees better.

Finally, some component trees have been improved. This was specifically
the case for the `Taetigkeit` component and its children. It also aligns
with the scheme we try to follow everywhere else.