fix(calculation): avoid out of memory issues by characterization tests
The characterization caused out of memory issues sometimes. This
happened specifically in the continuous integration environment. The
issue is that the tests creates all sample data at ones, resulting in
a huge array that can't be garbage collected. This got worse as the
resulting sample array was then mapped to the calculated results.
To avoid this issue, a loop is used to create one sample at a time and
verify the results one by one. Due to this, we end up with a different
snapshot file format. As nothing in the calculation code has been
touched, this is a safe operation.
We do not use a `*.each` test case to avoid bloating the test suite
without any benefit.