Revert commits 71b1e7cd and 54cf0b44
This is an artificial revert commit as reverting the commits the regular
way causes many conflicts and would result in multiple revert commits.

The development of the RadioInput component and the related work on the
CustomRadioGroup was not effective and caused some issues for new
features. Anyhow, a lot of research took place and will help us in
future development.