feat(components): better inputs for Gewinneinkünfte
Gewinneinkünfte and Lohneinkommen require different information by the
user. The calculation interface currently provides a unified model for
both as one. This is problematic. Especially because this leaked into
the user interface. Which led to users providing wrong inputs due to
a misleading visual interface.

Design has improved this and we have now alternative inputs for
Gewinneinkünfte. Under the hood it was also important to change the data
model of the form/store. Which then gets transformed into the model of
the calculation.

The whole code around the components etc. is highly copy-pasted and
still ugly and cumbersome. Anyhow, this is a small step into a better
world.