refactor: move out planer domain for stronger structure
This a first step of the intention to give the project a stronger and
more clean structure. A stronger emphasize on the directory structure
and naming should help. It is a potential intermediate step into further
structuring measurements.

It also starts to extend the code documentation. A first README local to
the new montasplaner package provides initial helpful information.