refactor: move out elterngeldrechner for stronger structure
This is a follow up on the intent to create a stronger and more clean
project directory structure. It finishes with the former suspicious
`globals/js` directory. A remnant of old migrations.

The naming of `elterngeldrechner` is a but unlucky, taking the
repository name itself into account. Though, the former `calculations`
does not properly fit too. But the repository (and the production
webiste) name isn't the best too. Therefore, this name is currently the
best for this package.