refactor: extract lebensmonatrechner into new package
Currently this logic is primarily used by the applications planer
feature. Thereby, it was located in the domain of the monatsplaner. But
this is only co-located by usage. Actually this a much more fundamental
logic which gets used by other features too. But for example the
elterngeldrechner package implements its own (wrong/buggy)
lebensmonatrechner too.
This should help to sharpen the focus of the different packages and
increase their re-usability.