refactor: extract lohnsteuerechner into new package
These algorithms are solely used by the elterngeldrechner. Anyhow, they
are kind of an dependency by the calculation. Historically this actually
was an external library. Before we had to replace it with a custom
implementation. Encapsulating this package provides some structure and
more overview in the elterngeldrechner. The latter now primarily focuses
on the implementation of the law and their guidelines.

The current approach uses a very strong/direct dependency model without
any inversion or injection. This might be some work for future
refactoring, nevertheless should be mentioned here.