refactor: extract user feedback as application feature
It might be very a big exaggerated to define this as a feature of the
user-interface. Anyhow, it is a collection of resources that as a whole
and should be free to move around. It should not matter where this user
feedback is actually integrated in the application. This is especially
true with our recent considerations to find better placements.
There is some dependency inversion to the tracking layer in place. This
is actually great, but feels a bit awkward in the context of the state
management in parallel to it. However this will be kept for now.

The code was kept close to the original. Though, some improvements based
on a good feature interface was made.

There were also minor improvements to some tests and code around the
Redux store management.