test: improve tests with Redux states
We already used `immer` for other tests, as it leads to much cleaner,
shorter and less messy tests (in context of Redux, which uses immer
itself).

Furthermore, the test utilities now expose the created store to further
simplify the usage in tests.

Moreover, the exported interface of the individual Redux slices has been
reduced to simplify dependencies. This will help future refactoring.