refactor: move state bounding into abfrageteil feature
There is no reason to expose this from the feature itself. It exposes
a much larger interface than actually necessary. This also means the
tests have shifted. Though, they basically stayed exactly the same.

Notice that the navigation issue isn't solved yet. The feature should
not know how to navigate where. But this is a broader issue in itself.