refactor: resolve component categories
This is an old design pattern that was still kept alive. Though, since
the project has moved the maintainer, it never felt just right anymore.
Anyhow, there were 0 organisms left, 2 atoms and a bunch of molecules.
Furthermore, the pages components never really fit into this scheme.

The new structure is not perfect, but should go into the right direct
and simplify things. The whole routing/pages part is not well refined
yet. But they have a stronger bonding than being re-usable components
for all features.

There are still some wild cross project dependencies that need to be
sorted. Some could be resolved with this commit.