refactor: extract routing from feature(s)
The features should be very independent and somewhat agnostic. After
all, they should be segregated from the routing logic. This
responsibility has to be pushed upwards.

What was special about this, were the usage of routing after forms got
submit. Furthermore, the submit/navigate button is outside the layout of
the form components of the abfrageteil feature. To solve this, an
external submit button is used. This was enabled by simple optional
properties for the forms. Thereby, testing remains stable and there is
no coupling to any routing logic at all.

Within this commit, the ButtonGroup component was resolved. This led to
kinda repetitive code. Anyhow, there is no 100% clear re-usable
structure yet. The costs are minimal.