refactor(monatsplaner): restructure passing of callbacks
The number of callbacks exposed by the monatsplaner feature starts to
grow. The former approach to use optional, positioned arguments does not
scale well. Therefore, this was restructured.

Furthermore, the naming was strengthened. Callbacks that are associated
1:1 with a use-case of the Planer service use the same naming now.
Though, the expressiveness of the `onPlanChanged` was reduced to
`onChange`. But it uses the same verb presence as typical for events and
it re-uses a commonly used term that everyone is familiar with. The
context of the monatsPLANER should do the rest.

Moreover, some tests have been added to actual test that the callbacks
are actually called. These have been partially missing.