feat(monatsplaner): add tracking for opening Lebensmonate
There is a hypotheses we need to verify. Therefore new data must be
tracked. As always: the new tracking variable will be picked up by a new
version of the tag manager.
The naming of the tracking variable and associated function, module, ...
is not great. Though, it nails it pretty much down to what is actually
is.

Unfortunately, after some refinement for what we actually need to track,
this is quite low level, down in the monatsplaner feature. We still
wanna keep the features focused and segregate responsibilities like
tracking. As the result, there is quite some "callback drilling" through
the component tree to forward the toggle event for the `<details>`
elements. As this event usually does not bubble by default, there is no
other possibility to observe this behavior. We hope this remains an
exception and might be removed soonish.