refactor: separate user-tracking related modules
The directory of `user-tracking` was quite unpleasant when opened. This
is partially due to the very expressive names of the metrics. But there
was also a lot of things mixed together all in one.
The new directory names should provide a little more guidance. Though,
the term `core` is a very generic word that is abused too often. With
further added complexity to it, its likely it should get separated from
the application and put into its own package. Only the application
specific metrics should then remain there. Potentially even aligned with
their related application features.