chore: convert StyleLint configuration to JavaScript
We try to uniformly use the emerging standard of tools in the JavaScript
space to use configuration files in plain JavaScript or even TypeScript
with naming like `<tool_name>.config.{js,ts}`. TypeScript is preferred,
but not supported by StyleLint yet.

This adds a few additional configuration options to report invalid usage
of comments to disable rules.