chore: remove obsolete Prettier plugin for TailwindCSS
Since we are using the `eslint-plugin-tailwindcss` this is no more
needed. The ESLint plugin does the same and can do more. Theoretically
a separation between linting and formatting is great. But unfortunately
not always possible.

In result the Prettier configuration was left empty. So it got deleted.
In future we would like to use the new file name and TypeScript (i.e.
`prettier.config.ts`).