feat: change some tracking in planer from variable to event
The variable tracking has the issue that it needs a trigger by the Tag
Manager to pick up the variable. This causes some trouble with the
quality of our tracking right now. Therefore, we try to use more
triggers from the code. The first one should be the opening of
Lebensmonate. The former count value of the variable can be derived in
the tracking system.