chore(user-tracking): remove obsolete metrics and merge by category
Due to the recent adaptions to use custom Matomo events instead of
variables, these metric modules have become obsolete. Pushing an event
is so straight forward, that its not worth to add any wrapper for this.
It creates the general question what the responsibility of the metric
modules should be in future.

Metrics get now fully sorted by their category action in the tag
manager. This was already partially done and has been unified now.