refactor(planer): clean up "callback" types of planer service
The decision to create this module was questionable from the beginning.
It didn't feel right, but it was somehow okayish. The new solution tries
to avoid a too central collection of type definitions and push them more
to the code locations where they are used. It doesn't feel perfect, but
its a little more "clean". At least somewhat.

This commit as the goal to clean up the `planer/service` directory to
allow for better extension to put more into it.