feat(planer): add logic to generate Beispiele
This is the first real puzzle piece for the new feature extension of the
Planer. Therefore it provides the data generation that will be then
visualized and interconnected into the Planer.

The code was on a "parking lot" because its exact location in the module
tree was unclear. The hooks directory of the feature are some ugly
escape mechanism to hide it. There is currently on clear and good
structure for such code.
The code was also developed with some challenges in mind that have been
discovered in successive steps (e.g. the identifiers). Thereby it might
not fully seem reasonable as a commit in itself. It was very hard to
develop it fully iterative.

While there are a good number of tests, there are TODO comments to add
important property test cases. This will need the planned data generator
functions.