feat(planer): add hook to provide Beispiele
This hook is intentionally separated from the planer service to maintain
responsibility and focus. The term "service" from a domain driven design
inspiration got inherited.
The hooks encapsulates all the relevant logic for this feature
extension, independent of the user interface. Though, it is known that
the logic would not fit into a single React component, as it must be
connected with the planer service. After all it provides nice
testability of all code units in the end.