feat(planer): tie up Beispiele into Planer
This simply connects all the puzzle pieces together. There is a minor
"tricky" part with some mutual dependency issue between the service
hooks of the feature.

Notice that we still don't have a good testing strategy for all of this.
The connection via callbacks between the services has no integration
test.